creditsense_ai/requirements.txt
pydantic>=2.0
python-dotenv
fastapi
uvicorn[standard]
python-multipart
google-generativeai
groq
openai
pymupdf
pdfplumber
pandas
networkx
gymnasium
numpy
scipy
web3
streamlit
websockets
python-docx
